Fix bugs.
You shouldn't use float for a int value.
Also you were trying to cast float value to float again.

+++ main.py
@@ -3,12 +3,12 @@ import os
 
 def main():
     maxdnm = 99999999
-    montant = float(input("Put Amount here: "))
-    dnm = float(input("Enter Coin Values, each one separated by a Space: "))
+    montant = int(input("Put Amount here: "))
+    dnm = input("Enter Coin Values, each one separated by a Space: ")
     dy = [maxdnm] * (montant + 1)
     input_list_str = dnm.split()
     dy[0] = 0
-    denominations = [float(item) for item in input_list_str]
+    denominations = [int(item) for item in input_list_str]
     for denominations in denominations:
         dy[denominations] = 1
     for i in range(1, montant + 1):
